fastapi/requirements.txt
fastapi==0.100.0
pydantic-settings==2.0.2
httpx==0.24.1
psycopg2-binary==2.9.6
pydantic==2.0.3
python-decouple==3.8
requests==2.31.0
SQLAlchemy==2.0.19
uvicorn==0.23.1
